Made apply_liger_kernel a private API
