+++ src/liger_kernel/transformers/__init__.py
@@ -1,5 +1,4 @@
 from liger_kernel.transformers.monkey_patch import (  # noqa: F401
-    apply_liger_kernel,
     apply_liger_kernel_to_gemma,
     apply_liger_kernel_to_llama,
     apply_liger_kernel_to_mistral,
